This adds GetZipBytes and GetTotalBytes for all containers appropriately.
What it does, is build up the sizes from the Container, the AuxContainer initially.. then goes ahead and fills it in with attributes and properties (other branches) and adds it to the owning container's size. This builds it up to get a dynamically-made sizes.pdf which shows sizes of the selection applied. This way, it doesn't hard-code anything and depends on the features a user puts in.

This is for #10.

What needs to be done still is add in extra features to build similar pie charts for Container instead of just for Type. This will be readily available soon since the code needs to be refactored cleverly.